Move File Operations to a Separate Module
* Move file ops to a separate module, export functions
* Remove Ava test
  * TODO: Replace with unit tests on new module
* Switch `require` -> `import`
