+++ index.js
@@ -1,50 +1,13 @@
 #!/usr/bin/env node
 
-const got = require('got');
-const inquirer = require('inquirer');
+import inquirer from 'inquirer';
 
-const API_URL = 'https://slack.com/api';
-
-const filterFiles = (files, options) => {
-  if (!files) {
-    console.warn(`There are no files to be deleted. Either there were no files older than ${options.daysOld} or there is something wrong with your token.`);
-    return [];
-  }
-
-  const removePinned = fileList => fileList.filter(file => !file.pinned_to);
-  const filesToDelete = options.keepPinned
-    ? removePinned(files)
-    : files;
-
-  if (!filesToDelete.length) {
-    console.warn('There are no files to be deleted.');
-  }
-
-  return filesToDelete;
-};
-
-const deleteFiles = (token, files) => {
-  console.log(`Deleting ${files.length} file(s)...`);
-
-  files.forEach(file => got(`${API_URL}/files.delete`, {
-    body: {
-      token,
-      file: file.id,
-    },
-  }).then(() => console.log(`${file.name} was deleted.`)).catch(error => console.error('Error while deleting files.', error)));
-};
+import { deleteFiles, filterFiles, getFiles } from './file_operations';
 
 const run = (token, options) => {
   const deleteOlderThan = Math.floor(new Date().getTime() / 1000) - (options.daysOld * 86400);
 
-  got(`${API_URL}/files.list`, {
-    body: {
-      token,
-      ts_to: deleteOlderThan,
-      count: 1000,
-    },
-    json: true,
-  })
+  getFiles(token, deleteOlderThan)
     .then(response => filterFiles(response.body.files, options))
     .then(files => deleteFiles(token, files))
     .catch(console.error);
